requirements.txt
streamlit
mysql-connector-python
cryptography
pandas
datetime
plotly